// web-application/src/types-openapi/api/models/UpdateBlogRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Badminton Court Management System API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface UpdateBlogRequest
 */
export interface UpdateBlogRequest {
    /**
     * 
     * @type {string}
     * @memberof UpdateBlogRequest
     */
    id?: string;
    /**
     * 
     * @type {string}
     * @memberof UpdateBlogRequest
     */
    title?: string | null;
    /**
     * 
     * @type {string}
     * @memberof UpdateBlogRequest
     */
    content?: string | null;
    /**
     * 
     * @type {string}
     * @memberof UpdateBlogRequest
     */
    imageUrl?: string | null;
    /**
     * 
     * @type {string}
     * @memberof UpdateBlogRequest
     */
    status?: string | null;
}

/**
 * Check if a given object implements the UpdateBlogRequest interface.
 */
export function instanceOfUpdateBlogRequest(value: object): value is UpdateBlogRequest {
    return true;
}

export function UpdateBlogRequestFromJSON(json: any): UpdateBlogRequest {
    return UpdateBlogRequestFromJSONTyped(json, false);
}

export function UpdateBlogRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): UpdateBlogRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'title': json['title'] == null ? undefined : json['title'],
        'content': json['content'] == null ? undefined : json['content'],
        'imageUrl': json['imageUrl'] == null ? undefined : json['imageUrl'],
        'status': json['status'] == null ? undefined : json['status'],
    };
}

export function UpdateBlogRequestToJSON(json: any): UpdateBlogRequest {
    return UpdateBlogRequestToJSONTyped(json, false);
}

export function UpdateBlogRequestToJSONTyped(value?: UpdateBlogRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'title': value['title'],
        'content': value['content'],
        'imageUrl': value['imageUrl'],
        'status': value['status'],
    };
}